tests/wpt/web-platform-tests/offscreen-canvas/the-offscreen-canvas/size.attributes.reflect.setidl.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>OffscreenCanvas test: size.attributes.reflect.setidl</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>

<h1>size.attributes.reflect.setidl</h1>
<p class="desc">Setting IDL attributes updates IDL and content attributes</p>


<script>
var t = async_test("Setting IDL attributes updates IDL and content attributes");
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

offscreenCanvas.width = 120;
offscreenCanvas.height = 60;
_assertSame(offscreenCanvas.width, 120, "offscreenCanvas.width", "120");
_assertSame(offscreenCanvas.height, 60, "offscreenCanvas.height", "60");

t.done();

});
</script>
